refactor(gateway): submodule per auth method
Move LDAP and anonymous authentication view in dedicated submodules,
under the new view.auth package.
